replace UI registry warning suffix
Add shorter @wallet-standard/ui-registry helper exports and move first-party packages and tests to those names.

Keep the warning-suffixed exports only as deprecated compatibility aliases, and rename UI registry modules and tests so generated package filenames no longer carry the suffix.

Closes #133
