+++ .changeset/modern-ui-wallet-registry-helper.md
@@ -1,0 +1,8 @@
+---
+"@wallet-standard/react-core": patch
+"@wallet-standard/ui-compare": patch
+"@wallet-standard/ui-features": patch
+"@wallet-standard/ui-registry": minor
+---
+
+Add shorter UI registry helper exports, retain the warning-suffixed exports as deprecated aliases, remove the warning suffix from UI registry module filenames, and use the shorter helpers in first-party packages.
