+++ packages/react/core/src/useWallets.ts
@@ -1,5 +1,5 @@
 import type { UiWallet } from '@wallet-standard/ui';
-import { getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from '@wallet-standard/ui-registry';
+import { getOrCreateUiWalletForStandardWallet } from '@wallet-standard/ui-registry';
 import { useMemo } from 'react';
 
 import { useWallets_INTERNAL_ONLY_NOT_FOR_EXPORT } from './useWallets_INTERNAL_ONLY_NOT_FOR_EXPORT.js';
@@ -9,9 +9,6 @@ import { useWallets_INTERNAL_ONLY_NOT_FOR_EXPORT } from './useWallets_INTERNAL_O
  */
 export function useWallets(): readonly UiWallet[] {
     const wallets = useWallets_INTERNAL_ONLY_NOT_FOR_EXPORT();
-    const uiWallets = useMemo(
-        () => wallets.map(getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_FIRED),
-        [wallets]
-    );
+    const uiWallets = useMemo(() => wallets.map(getOrCreateUiWalletForStandardWallet), [wallets]);
     return uiWallets;
 }
