+++ packages/ui/compare/src/__tests__/compare-test.ts
@@ -1,8 +1,8 @@
 import type { Wallet, WalletAccount, WalletVersion } from '@wallet-standard/base';
 import type { UiWalletAccount, UiWallet } from '@wallet-standard/ui-core';
 import {
-    getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
-    getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_FIRED,
+    getOrCreateUiWalletAccountForStandardWalletAccount,
+    getOrCreateUiWalletForStandardWallet,
 } from '@wallet-standard/ui-registry';
 
 import { uiWalletAccountBelongsToUiWallet, uiWalletAccountsAreSame } from '../compare.js';
@@ -26,36 +26,30 @@ describe('uiWalletAccountsAreSame()', () => {
             name: 'Mock:Wallet',
             version: '1.0.0' as WalletVersion,
         };
-        mockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(
-            mockWallet,
-            mockWalletAccount
-        );
+        mockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(mockWallet, mockWalletAccount);
     });
     it('returns true for the same account object', () => {
         expect(uiWalletAccountsAreSame(mockUiWalletAccount, mockUiWalletAccount)).toBe(true);
     });
     it('returns true if the addresses and underlying `Wallet` match, despite the objects being different', () => {
-        const clonedMockUiWalletAccount =
-            getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(mockWallet, {
-                ...mockWalletAccount,
-                chains: ['solana:danknet'],
-            } as WalletAccount);
+        const clonedMockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(mockWallet, {
+            ...mockWalletAccount,
+            chains: ['solana:danknet'],
+        } as WalletAccount);
         expect(uiWalletAccountsAreSame(mockUiWalletAccount, clonedMockUiWalletAccount)).toBe(true);
     });
     it('returns false if the addresses match but the underlying `Wallet` does not', () => {
-        const clonedMockUiWalletAccount =
-            getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(
-                { ...mockWallet },
-                mockWalletAccount
-            );
+        const clonedMockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(
+            { ...mockWallet },
+            mockWalletAccount
+        );
         expect(uiWalletAccountsAreSame(mockUiWalletAccount, clonedMockUiWalletAccount)).toBe(false);
     });
     it('returns false if the addresses do not match even if the underlying `Wallet` does', () => {
-        const clonedMockUiWalletAccount =
-            getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(mockWallet, {
-                ...mockWalletAccount,
-                address: 'xyz',
-            });
+        const clonedMockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(mockWallet, {
+            ...mockWalletAccount,
+            address: 'xyz',
+        });
         expect(uiWalletAccountsAreSame(mockUiWalletAccount, clonedMockUiWalletAccount)).toBe(false);
     });
 });
@@ -80,22 +74,18 @@ describe('uiWalletAccountBelongsToUiWallet()', () => {
             name: 'Mock:Wallet',
             version: '1.0.0' as WalletVersion,
         };
-        mockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(
-            mockWallet,
-            mockWalletAccount
-        );
-        mockUiWallet = getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(mockWallet);
+        mockUiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(mockWallet, mockWalletAccount);
+        mockUiWallet = getOrCreateUiWalletForStandardWallet(mockWallet);
     });
     it('returns true if the UI wallet account belongs to the UI wallet', () => {
         expect(uiWalletAccountBelongsToUiWallet(mockUiWalletAccount, mockUiWallet)).toBe(true);
     });
     it('returns false if the UI wallet account does not belong to the UI wallet', () => {
         const differentWallet = { ...mockWallet } as Wallet; /* different object */
-        const identicalAccountInDifferentWallet =
-            getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(
-                differentWallet,
-                mockWalletAccount
-            );
+        const identicalAccountInDifferentWallet = getOrCreateUiWalletAccountForStandardWalletAccount(
+            differentWallet,
+            mockWalletAccount
+        );
         expect(uiWalletAccountBelongsToUiWallet(identicalAccountInDifferentWallet, mockUiWallet)).toBe(false);
     });
 });
