+++ packages/ui/compare/src/storage-key.ts
@@ -1,5 +1,5 @@
 import type { UiWalletAccount } from '@wallet-standard/ui-core';
-import { getWalletForHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from '@wallet-standard/ui-registry';
+import { getWalletForHandle } from '@wallet-standard/ui-registry';
 
 /**
  * Produces a stable string that can be used to uniquely identify a `UiWalletAccount`.
@@ -8,6 +8,6 @@ import { getWalletForHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from '@wallet-stan
  * your app's last selected account as a string in browser storage.
  */
 export function getUiWalletAccountStorageKey(uiWalletAccount: UiWalletAccount): string {
-    const underlyingWallet = getWalletForHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(uiWalletAccount);
+    const underlyingWallet = getWalletForHandle(uiWalletAccount);
     return `${underlyingWallet.name.replace(':', '_')}:${uiWalletAccount.address}`;
 }
