+++ packages/ui/features/src/getWalletFeature.ts
@@ -4,7 +4,7 @@ import {
     safeCaptureStackTrace,
 } from '@wallet-standard/errors';
 import type { UiWalletHandle } from '@wallet-standard/ui-core';
-import { getWalletForHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from '@wallet-standard/ui-registry';
+import { getWalletForHandle } from '@wallet-standard/ui-registry';
 
 /**
  * Returns the feature object from the Wallet Standard `Wallet` that underlies a `UiWalletHandle`.
@@ -14,7 +14,7 @@ export function getWalletFeature<TWalletHandle extends UiWalletHandle>(
     uiWalletHandle: TWalletHandle,
     featureName: TWalletHandle['features'][number]
 ): unknown {
-    const wallet = getWalletForHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(uiWalletHandle);
+    const wallet = getWalletForHandle(uiWalletHandle);
     if (!(featureName in wallet.features)) {
         const err = new WalletStandardError(WALLET_STANDARD_ERROR__FEATURES__WALLET_FEATURE_UNIMPLEMENTED, {
             featureName,
