+++ packages/ui/registry/src/UiWalletRegistry.ts
@@ -1,8 +1,8 @@
 import type { IdentifierArray, Wallet } from '@wallet-standard/base';
 import type { UiWalletAccount, UiWallet } from '@wallet-standard/ui-core';
 
-import { getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from './UiWalletAccountRegistry_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.js';
-import { registerWalletHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED } from './UiWalletHandleRegistry_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.js';
+import { getOrCreateUiWalletAccountForStandardWalletAccount } from './UiWalletAccountRegistry.js';
+import { registerWalletHandle } from './UiWalletHandleRegistry.js';
 import { identifierArraysAreDifferent } from './compare.js';
 
 const walletsToUiWallets = new WeakMap<Wallet, UiWallet>();
@@ -10,16 +10,12 @@ const walletsToUiWallets = new WeakMap<Wallet, UiWallet>();
 type Mutable<T> = { -readonly [P in keyof T]: T[P] };
 
 /**
- * DO NOT USE THIS OR YOU WILL BE FIRED
+ * Create or obtain the existing `UiWallet` object associated with a Wallet Standard `Wallet`.
  *
- * This method is for exclusive use by Wallet Standard UI library authors. Use this if you need to
- * create or obtain the existing `UiWallet` object associated with a Wallet Standard `Wallet`.
- *
- * @internal
+ * This method is intended for Wallet Standard UI library authors, or app authors building directly
+ * on the UI primitives without a framework-specific adapter.
  */
-export function getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_FIRED<TWallet extends Wallet>(
-    wallet: TWallet
-): UiWallet {
+export function getOrCreateUiWalletForStandardWallet<TWallet extends Wallet>(wallet: TWallet): UiWallet {
     const existingUiWallet = walletsToUiWallets.get(wallet);
     const mustInitialize = !existingUiWallet;
     let uiWallet: Mutable<UiWallet> = existingUiWallet ?? ({} as Mutable<UiWallet>);
@@ -37,11 +33,7 @@ export function getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_F
                 yield* this._cache;
             }
             for (const walletAccount of wallet.accounts.slice(this._cache.length)) {
-                const uiWalletAccount =
-                    getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(
-                        wallet,
-                        walletAccount
-                    );
+                const uiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(wallet, walletAccount);
                 this._cache.push(uiWalletAccount);
                 yield uiWalletAccount;
             }
@@ -51,11 +43,7 @@ export function getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_F
                 return true;
             }
             for (const walletAccount of wallet.accounts.slice(this._cache.length)) {
-                const uiWalletAccount =
-                    getOrCreateUiWalletAccountForStandardWalletAccount_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(
-                        wallet,
-                        walletAccount
-                    );
+                const uiWalletAccount = getOrCreateUiWalletAccountForStandardWalletAccount(wallet, walletAccount);
                 this._cache.push(uiWalletAccount);
                 if (predicateFn(uiWalletAccount)) {
                     return true;
@@ -99,7 +87,11 @@ export function getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_F
     }
     if (isDirty) {
         walletsToUiWallets.set(wallet, uiWallet);
-        registerWalletHandle_DO_NOT_USE_OR_YOU_WILL_BE_FIRED(uiWallet, wallet);
+        registerWalletHandle(uiWallet, wallet);
     }
     return Object.freeze(uiWallet);
 }
+
+/** @deprecated Use {@link getOrCreateUiWalletForStandardWallet} instead. */
+export const getOrCreateUiWalletForStandardWallet_DO_NOT_USE_OR_YOU_WILL_BE_FIRED =
+    getOrCreateUiWalletForStandardWallet;
