+++ packages/ui/registry/src/index.ts
@@ -1,3 +1,3 @@
-export * from './UiWalletAccountRegistry_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.js';
-export * from './UiWalletHandleRegistry_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.js';
-export * from './UiWalletRegistry_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.js';
+export * from './UiWalletAccountRegistry.js';
+export * from './UiWalletHandleRegistry.js';
+export * from './UiWalletRegistry.js';
